Add mezcal page
