+++ pages/tequila.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Head from "next/head";
+//import Head from "next/head";
 //import Image from "next/image";
 //import Icon from "../public/Assets/Agave.jpg"
 //import TequilaCard from "../components/TequilaCard";
